Move to larger int field

+++ wcivf/apps/people/models.py
@@ -20,7 +20,7 @@ class PersonPost(models.Model):
     election = models.ForeignKey(Election, null=False, on_delete=models.CASCADE)
     list_position = models.IntegerField(blank=True, null=True)
     elected = models.NullBooleanField()
-    votes_cast = models.PositiveSmallIntegerField(null=True)
+    votes_cast = models.PositiveIntegerField(null=True)
     objects = PersonPostManager()
 
     def __str__(self):
